+++ lib/byebug/commands/save.rb
@@ -54,7 +54,7 @@ module Byebug
     end
 
     def save_settings(file)
-      %w(autoeval autoirb autolist basename).each do |setting|
+      %w(autoirb autolist basename).each do |setting|
         file.puts "set #{setting} #{Setting[setting.to_sym]}"
       end
     end