+++ lib/byebug/helpers/var.rb
@@ -1,21 +1,16 @@
+require 'byebug/helpers/eval'
+
 module Byebug
   module Helpers
     #
     # Utilities for variable subcommands
     #
     module VarHelper
-      def var_list(ary, b = get_binding)
-        vars = ary.sort.map do |v|
-          s = begin
-            b.eval(v.to_s).inspect
-          rescue
-            begin
-              b.eval(v.to_s).to_s
-            rescue
-              '*Error in evaluation*'
-            end
-          end
-          [v, s]
+      include EvalHelper
+
+      def var_list(ary, binding = default_binding)
+        vars = ary.sort.map do |name|
+          [name, safe_inspect(silent_eval(name.to_s, binding))]
         end
         puts prv(vars, 'instance')
       end
@@ -29,7 +24,7 @@ module Byebug
       end
 
       def var_instance(str)
-        obj = bb_warning_eval(str || 'self')
+        obj = single_thread_eval(str || 'self')
 
         var_list(obj.instance_variables, obj.instance_eval { binding })
       end
@@ -39,6 +34,18 @@ module Byebug
         cur_self = @state.context.frame_self(@state.frame)
         locals[:self] = cur_self unless cur_self.to_s == 'main'
         puts prv(locals.keys.sort.map { |k| [k, locals[k]] }, 'instance')
+      end
+
+      def safe_inspect(var)
+        var.inspect
+      rescue
+        safe_to_s(var)
+      end
+
+      def safe_to_s(var)
+        var.to_s
+      rescue
+        '*Error in evaluation*'
       end
     end
   end