+++ test/commands/show_test.rb
@@ -5,8 +5,8 @@ module Byebug
   # Test settings display functionality.
   #
   class ShowTestCase < TestCase
-    settings = %i(autoeval autolist autosave basename fullpath post_mortem
-                  stack_on_error)
+    settings =
+      %i(autolist autosave basename fullpath post_mortem stack_on_error)
 
     settings.each do |set|
       define_method(:"test_show_#{set}_shows_disabled_bool_setting_#{set}") do